Add multi client connection capability, restructure project and delete unnecessary files(test files)
