--- Socket Server/test.java
@@ -1,1 +1,0 @@
-System.out.println("Test")
